floss: Fix clippy warnings in topshim
- warning: an implementation of `From` is preferred since it gives you
  `Into<_>` for free where the reverse isn't true
- warning: consider choosing a more descriptive name
- warning: deref on an immutable reference
- warning: digits of hex, binary or octal literal not in groups of
  equal size
- warning: doc list item overindented
- warning: found `.or(Some(…)).unwrap()`
- warning: large size difference between variants
- warning: manual implementation of `.is_multiple_of()`
- warning: method `default` can be confused for the standard trait
  method `std::default::Default::default`
- warning: the following explicit lifetimes could be elided:
- warning: this `impl` can be derived
- warning: this expression creates a reference which is immediately
  dereferenced by the compiler
- warning: this function has too many arguments
- warning: unneeded `return` statement
- warning: unsafe function's docs are missing a `# Safety` section
- warning: use of `offset` with a `usize` casted to an `isize`
- warning: using `clone` on type `*const Uuid` which implements the
  `Copy` trait
- warning: very complex type used. Consider factoring parts into `type`
  definitions
- warning: you should consider adding a `Default` implementation for

Bug: 343315863
Test: m com.google.android.bt
Test: floss/build/build-in-container.py
Test: build.py --target clippy
Test: emerge-cherry floss-upstream
Test: bluetooth_AdapterSAHealth.all_floss
Flag: NONE Floss only change

+++ floss/rust/topshim/src/metrics.rs
@@ -23,6 +23,7 @@ mod ffi {
             bond_state: u32,
             fail_reason: i32,
         );
+        #[allow(clippy::too_many_arguments)]
         fn device_info_report(
             bt_addr: RawAddress,
             device_type: u32,
@@ -73,10 +74,11 @@ pub fn bond_state_changed(
         device_type as u32,
         status as u32,
         bond_state as u32,
-        fail_reason as i32,
+        fail_reason,
     );
 }
 
+#[allow(clippy::too_many_arguments)]
 pub fn device_info_report(
     addr: RawAddress,
     device_type: BtDeviceType,
@@ -90,7 +92,7 @@ pub fn device_info_report(
     ffi::device_info_report(
         addr,
         device_type as u32,
-        class_of_device as u32,
+        class_of_device,
         appearance as u32,
         vendor_id as u32,
         vendor_id_src as u32,
